Add support for summary statistics.
The summary statistics table is very small, and is intended
to be populated via cron to avoid any lag in the actual webapp
when scanning all of the rows in players, games, and servers.
Even caching such a query would yield bad performance during
invalidation, so I've opted to make this table be populated with
an offline script similar to the one for ranks.